Fix constructing i18n directory
`this.app.trees.app` can be a string pointing to the app directory or a
broccoli node. In cases where it is a broccoli node, performing a string
concat will not give the intended result. `this.project.root` should
always be a string, as that is what it is supposed to be doing.